MCR-2909 move code from MCRObjectID to MCRObjectIDGenerator
