+++ mycore-base/src/test/java/org/mycore/datamodel/metadata/MCRObjectIDTest.java
@@ -35,6 +35,7 @@ import org.jdom2.Element;
 import org.junit.Before;
 import org.junit.Test;
 import org.mycore.common.MCRStoreTestCase;
+import org.mycore.datamodel.common.MCRObjectIDGenerator;
 
 public class MCRObjectIDTest extends MCRStoreTestCase {
 
@@ -48,12 +49,12 @@ public class MCRObjectIDTest extends MCRStoreTestCase {
 
     @Test
     public void setNextFreeIdString() {
-        MCRObjectID id1 = MCRObjectID.getNextFreeId(BASE_ID);
+        MCRObjectID id1 = MCRObjectIDGenerator.getNextFreeId(BASE_ID);
         assertEquals("First id should be int 1", 1, id1.getNumberAsInteger());
-        MCRObjectID id2 = MCRObjectID.getNextFreeId(BASE_ID);
+        MCRObjectID id2 = MCRObjectIDGenerator.getNextFreeId(BASE_ID);
         assertEquals("Second id should be int 2", 2, id2.getNumberAsInteger());
         getStore().create(id2, new Document(new Element("test")), new Date());
-        MCRObjectID id3 = MCRObjectID.getNextFreeId(BASE_ID);
+        MCRObjectID id3 = MCRObjectIDGenerator.getNextFreeId(BASE_ID);
         assertEquals("Second id should be int 3", 3, id3.getNumberAsInteger());
     }
